+++ src/main/scala/core/layout/utils/Locator.scala
@@ -1,6 +1,6 @@
 package core.layout.utils
 
-import java.awt.Point
+import core.shape.helper.Point
 
 import core.component.{BasicComponent, VisualPanel}
 import core.layout.LayoutManager.LayerManager
@@ -12,8 +12,8 @@ trait Locator {
   def validate(visualPanel: VisualPanel, layerID: Int): Boolean =
     LayerManager.validate(this)(visualPanel, layerID)
 
-  def getX: Int
-  def getY: Int
+  def getX: Float
+  def getY: Float
   def getPoint: Point = new Point(getX, getY)
 
 }
